Rename and cleanup - keeping jupyter shims
Given this is not jupyter-centric anymore, we rename everything as just widget
