+++ tsconfig.json
@@ -13,7 +13,7 @@
   "include": [
     "src/**/*",
     "app/**/*",
-    "python/jupyter/src/**/*",
+    "python/widget/src/**/*",
     "python/streamlit/src/**/*"
   ],
   "exclude": [
